@@ -1,6 +1,7 @@
 package com.pax.helloworld;
 
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
 
 public class AppInfo {
     /**
