@@ -1,7 +1,8 @@
 package com.pax.helloworld.second;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.pax.helloworld.R;
 
