@@ -1,9 +1,10 @@
 package com.pax.helloworld.third;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.pax.helloworld.LogUtil;
 import com.pax.helloworld.R;
